Split Documentation.cs
